Introduce optimized recursive Graph-Stmt
Use a modified Version of the graph-data
query statement. Instead of looping
through by ourselves we use the PostgreSQL
recursive WITH Query to let the database do this for us

This increases the initial cost of the query but it scales
much more nicely when performing queries on much more
complex data or with a much bigger depth.