add git group to config and manually chown
There seems to be an error inside of
pwd.c/h which has a function defined, that
is called by FileUtils.chown_R, which should
determine the uid based on the username, but fails
to do so (at least on Mac OS X 10.7.5).
As substitute we call chown via a system call.