do not rely on angle brackets anymore to ident uri
Hets does not use angle brackets anymore, so we do not
either. Now we try to determine for ourselves
if the supplied name (usually comes from the name
attribute of a DGNode in hets xml output) is a url
and if we should convert it to a, more readable,
name.