support references in distributed ontologies
distributed ontologies (actually single ontologies in distributed
ontologies) can also import and reference ontologies which are not part
of the distributed ontology. So we cannot use `iri_for_child` to infer
the ontology-iri on ontohub for a referenced DGNode name. Instead we
create an ontology aliasing map which maps DGNode names to their
corresponding ontology-iri, which is determined when creating/finding
the specific ontologies.

This leverages the fact that DGLinks are always placed after the
DGNodes in hets-xml output.