adjust hets version checking schema
Adjust the hets-version checking to check for a version string like
this: v0.99, 1402678179

Also change the way the method complains about non-matching versions.
Versions which are are lesser than (through lexicographical comparison)
the target string will still raise the VersionOutdated error but
versions which fall through the regex-check (non-compatible) will only
produce warnings in the log and on STDERR instead of killing the
application.