Ajaxify directory creation
The 'Create subdirectory' button now adds a line to the file browser
containing a form. When this form is sent, an XHR is sent. If the
directory could be created, the form is replaced by a link. Otherwise
an error message is shown.

The user can show as many forms as he wants.