adjust presence-validation for tempfile/iri
As only one of the two needs to be provided, we had to change the
validations (and add a validation for the remote file iri).
Additionally, as simple-form and html5 do not support conditionally
required fields we shall not supply an option for that.
Local file upload is marked as default for the radio-button mode