provide server-options for hets and use -a none
Inside of the hets.yml there is a key server_options which
will allow us to pass arguments to the hets-server instance which
is managed by god or by the rake task (which is in turn called by
invoker in development).

As an additional gimmick: The rake task for hets now uses
the search-paths in the hets.yml.