explicitly set serializer for ontologies#show
The reason that this is necessary is because of the fact that we
actually do not have Ontology per se, but DistributedOntology and
SingleOntology, so ActiveMode::Serializers would look for
SingleOntologySerializer or DistributedOntologySerializer. Instead of
providing both of them (with the same implementation), we'll just
declare that it should use the OntologySerializer.