Process ProveOptions and their ProofAttempts in CollectiveProofAttempt
Proving and importing the proof now takes place in the class
CollectiveProofAttempt instead of Theorem and OntologyVersion.

Also the ProveOptions are now prepared synchronously in the class Proof
and mapped to the list of ProofAttempts that will use those options.

When importing the proofs from Hets, this mapping will be needed.