Use config.fqdn instead of Settings.hostname
If `Settings.hostname` is defined, it is the same as `config.fqdn`,
so this is only really a change if `Settings.hostname` is `nil`.

This doesn't change the deployment config. There we explicitly use
`Settings.hostname` because we don't want to read the hostname from our
development machines when deploying.