Fix ssh key fingerprint generation
Openssh changed the fingerprint output format in version 6.8. Now it
defaults to SHA256 (previously MD5) and it prepends `MD5:` or `SHA256:`
to the fingerprint.

With `-E md5` we tell it to use the old fingerprint format. This option
is not recognized by versions before 6.8 and calling `ssh-keygen` with
it lets the process fail. To support the older versions as well
(they are running on our productive deployments), we call it a second
time with correct options if we find ourselves in in an environment with
the newer version.