send resource as argument to retry_resource_chain
Helper-Methods are not evaluated in the scope of the view (with its
local variables) but instead in the scope of the main-view/controller.
So resource would evaluate to the inherited-resource method instead of
the local variable inside of the _state-view. To remove the problem, we
inject resource as argument to the method.