Allow to provers data migration to fail for ontology versions
a) Hets closes the HTTP connection after a minute, which results in
timeouts. If a timeout occurs, we try again to get the provers list
twice.

b) Sometimes another error occurs (Hets response not usable).

In both cases we print the ID of the OntologVersion to know which ones
fail. This way, we can process them in another data migration.