Remove method that overwrites sidekiq options
The removed comment above the method is not only misleading but also
wrong. get_sidekiq_options does not need to be defined here because it
is defined in the included module Sidekiq::Worker. It returns what was
defined by sidekiq_options.

This method overwrote the sidekiq_options which set the queue to 'hets'.