Refactor all the Worker classes for parsing
Now, all comes down to the new OntologyParsingWorker. It performs on a
queue of triples containing the version id, options for the version and
the try_count (number of the next parsing attempt).

If parsing a version fails, it is put back to the queue (at the end),
with an incremented try_count. If this happens too often for the same
version, it is given to the SequentialOntologyParsingWorker that only
parses one OntologyVersion at a time.

The OntologyParsingPriorityWorker is a more elegant way to use the
priotitized queue.

All the background parsing is now performed via the
OntologyParsingWorker or its subclasses.