Fix ontology spec on error while parsing
The ontology needs to be reloaded because sidekiq does so as well. It
always fetches a fresh record from the database.
The test failed before because `NodeEvaluationHelper#clean_ontology`
called destroy on all the ontology's symbols which existed in the
object, but were not really persisted because of an aborted
transaction.